refactor(ipc): centralize api contract
