+++ tsconfig.web.json
@@ -4,6 +4,7 @@
     "src/renderer/src/env.d.ts",
     "src/renderer/src/**/*",
     "src/renderer/src/**/*.tsx",
+    "src/preload/api-types.ts",
     "src/preload/*.d.ts"
   ],
   "compilerOptions": {
